User update functionality added, UserNotFoundException created

+++ src/main/java/com/bae/service/UserService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.bae.exceptions.UserNotFoundException;
 import com.bae.persistence.domain.User;
 import com.bae.persistence.repo.UserRepo;
 
@@ -27,8 +28,24 @@ public class UserService {
 	}
 	
 	public String deleteUser(Long userId) {
+		if (!userRepo.existsById(userId)) {
+			throw new UserNotFoundException();
+		}
 		userRepo.deleteById(userId);
 		return "User deleted";
+	}
+	
+	public User findUserById(Long userId) {
+		return userRepo.findById(userId).orElseThrow(
+				() -> new UserNotFoundException());
+	}
+	
+	public User updateUser(User user, Long id) {
+		User toUpdate = findUserById(id);
+		toUpdate.setUsername(user.getUsername());
+		toUpdate.setPassword(user.getPassword());
+		toUpdate.setBoulders(user.getBoulders());
+		return this.userRepo.save(toUpdate);
 	}
 
 }
